+++ TransnetAndroid/app/src/main/java/whitewire/transnetandroid/MainActivity.java
@@ -1,466 +1,52 @@
 package whitewire.transnetandroid;
 
-import android.database.Cursor;
-import android.provider.MediaStore;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.content.Intent;
-import android.annotation.TargetApi;
-import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.preference.PreferenceManager;
-import android.support.design.widget.Snackbar;
-import android.support.annotation.NonNull;
-import android.app.LoaderManager.LoaderCallbacks;
-
-import android.content.CursorLoader;
-import android.content.Loader;
-import android.net.Uri;
-import android.os.AsyncTask;
-import android.Manifest;
-
-import android.os.Build;
 import android.os.Bundle;
-import android.util.Base64;
+import android.provider.MediaStore;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
-import android.provider.ContactsContract;
-import android.text.TextUtils;
-import android.view.KeyEvent;
-import android.view.View.OnClickListener;
-import android.view.inputmethod.EditorInfo;
-import android.widget.ArrayAdapter;
-import android.widget.AutoCompleteTextView;
-import android.widget.EditText;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
 
-import com.android.volley.AuthFailureError;
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.StringRequest;
-import com.android.volley.toolbox.Volley;
+import com.google.android.gms.wallet.OfferWalletObject;
 
-import java.io.ByteArrayOutputStream;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import static android.Manifest.permission.READ_CONTACTS;
+import static whitewire.transnetandroid.R.id.btnCamera;
 
 /**
- * A login screen that offers login via email/password.
+ * Created by Claudio on 25-Feb-18.
  */
-public class MainActivity extends AppCompatActivity implements LoaderCallbacks<Cursor> {
 
-    /**
-     * Id to identity READ_CONTACTS permission request.
-     */
-    private static final int REQUEST_READ_CONTACTS = 0;
+public class MainActivity extends AppCompatActivity{
+    public void checkPreferences () {
+        Boolean isFirstRun = getSharedPreferences("PREFERENCE", MODE_PRIVATE)
+                .getBoolean("isFirstRun", true);
 
-    /**
-     * A dummy authentication store containing known user names and passwords.
-     * TODO: remove after connecting to a real authentication system.
-     */
-    private static final String[] DUMMY_CREDENTIALS = new String[]{
-            "foo@example.com:hello", "bar@example.com:world"
-    };
-    /**
-     * Keep track of the login task to ensure we can cancel it if requested.
-     */
-    private UserLoginTask mAuthTask = null;
-
-    private SharedPreferences sharedPreferences;
-
-    // UI references.
-    private AutoCompleteTextView mEmailView;
-    private EditText mPasswordView, mCreditView;
-    private ImageView mBitmapView;
-
-    ImageView imageView;
-    Bitmap bitmap;
-
-    private String mEmail, mPassword, mCredit, mUrl = null;
-
-    protected void checkPreferences() {
-        boolean firstRun = getSharedPreferences("PREFERENCES", MODE_PRIVATE)
-                .getBoolean("firstRun", true);
-
-        if (!firstRun) {
-            Intent intent_new = new Intent(this, RideSelectionActivity.class);
-            startActivity(intent_new);
+        // If the user is logging in for the first time, show them the starting menu
+        if (isFirstRun) {
+            Intent intent = new Intent(this, RegisterActivity.class);
+            startActivity(intent);
         }
     }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        setContentView(R.layout.ride_selection);
         checkPreferences();
 
-        setContentView(R.layout.activity_main);
-        // boolean isFirstRun = getSharedPreferences("PREFERENCES", MODE_PRIVATE)
-        //         .getBoolean("firstRun", true);
-        sharedPreferences = PreferenceManager.getDefaultSharedPreferences(MainActivity.this);
-        SharedPreferences.Editor edit = sharedPreferences.edit();
-
-        // Initialize the firstRun and imageUploaded characteristics
-        // edit.putBoolean("firstRun", false).apply();
-        // edit.putBoolean("imageUploaded", false).apply();
-
-        // Set up the login form.
-        mEmailView = (AutoCompleteTextView) findViewById(R.id.email);
-        populateAutoComplete();
-
-        mCreditView = (EditText) findViewById(R.id.credit);
-
-        mPasswordView = (EditText) findViewById(R.id.password);
-        mPasswordView.setOnEditorActionListener(new TextView.OnEditorActionListener() {
-            @Override
-            public boolean onEditorAction(TextView textView, int id, KeyEvent keyEvent) {
-                if (id == EditorInfo.IME_ACTION_DONE || id == EditorInfo.IME_NULL) {
-                    attemptLogin();
-                    return true;
-                }
-                return false;
-            }
-        });
-
-        final Button registerButton = (Button) findViewById(R.id.register);
-        final Button mSendPictureButton = (Button) findViewById(R.id.send_picture_button);
-        Button btnCamera = (Button) findViewById(R.id.btnCamera);
-        imageView = (ImageView) findViewById(R.id.imageView);
-
-        mSendPictureButton.setEnabled(false);
-        registerButton.setEnabled(false);
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            if (checkSelfPermission(Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-                ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.CAMERA}, 1);
-            }
-        }
-        btnCamera.setOnClickListener(new View.OnClickListener() {
+        final Button requireRide = (Button) findViewById(R.id.requireRide);
+        final Button offerRide = (Button) findViewById(R.id.offerRide);
+        requireRide.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                startActivityForResult(intent, 0);
-                mSendPictureButton.setEnabled(true);
+                startActivity(new Intent(getApplicationContext(), MapsActivity.class));
             }
         });
 
-        mSendPictureButton.setOnClickListener(new OnClickListener() {
+        offerRide.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                attemptLogin();
-                if (mEmail != null && mPassword != null && mCredit != null) {
-                    registerButton.setEnabled(true);
-                    ByteArrayOutputStream stream = new ByteArrayOutputStream();
-                    bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream);
-                    byte[] byteArray = stream.toByteArray();
-                    mUrl = Base64.encodeToString(byteArray, Base64.DEFAULT);
-                } else {
-                    Toast.makeText(getApplicationContext(), "Verify credentials",
-                        Toast.LENGTH_LONG).show();
-                }
+                startActivity(new Intent(getApplicationContext(), OfferRideActivity.class));
             }
         });
-
-        registerButton.setOnClickListener(new OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                // Set preferences to say the user is already logged in
-                getSharedPreferences("PREFERENCES", MODE_PRIVATE).edit()
-                        .putBoolean("firstRide", false);
-
-                // Setting up request
-                RequestQueue mRequestQueue = Volley.newRequestQueue(getApplicationContext());
-
-                // Start the queue
-                mRequestQueue.start();
-                String url = "http://api-transnet.azurewebsites.net/api/Values/Register";
-
-                // Formulate the request and handle the response.
-                StringRequest stringRequest = new StringRequest(Request.Method.POST, url,
-                        new Response.Listener<String>() {
-                            @Override
-                            public void onResponse(String response) {
-                                if (response == "false") {
-                                    Toast.makeText(getApplicationContext(), "Registration unsuccessful.\n" +
-                                            "Please verify your information.",
-                                            Toast.LENGTH_LONG).show();
-                                } else {
-                                    startActivity(new Intent(getApplicationContext(), RideSelectionActivity.class));
-                                }
-                            }
-                        },
-                        new Response.ErrorListener() {
-                            @Override
-                            public void onErrorResponse(VolleyError error) {
-                                Toast.makeText(getApplicationContext(), error.toString(),
-                                        Toast.LENGTH_LONG).show();
-                            }
-                        }) {
-
-                // Sending bounds to the API to return value for given limits
-                @Override
-                protected Map<String, String> getParams() {
-                    Map<String, String> parameters = new HashMap<>();
-                    parameters.put("Email", mEmail);
-                    parameters.put("Password", mPassword);
-                    parameters.put("ImageUrl", mUrl);
-                    parameters.put("payment_no", mCredit);
-                    return parameters;
-                }
-
-                    @Override
-                    public Map<String, String> getHeaders() throws AuthFailureError {
-                        Map<String, String> params = new HashMap<>();
-                        return params;
-                    }
-                };
-
-                // Add the request to the RequestQueue.
-                mRequestQueue.add(stringRequest);
-            }
-        });
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        //Log.d("Data:", data.toString());
-        super.onActivityResult(requestCode, resultCode, data);
-        bitmap = (Bitmap) data.getExtras().get("data");
-        imageView.setImageBitmap(bitmap);
-    }
-
-
-    private void populateAutoComplete() {
-        if (!mayRequestContacts()) {
-            return;
-        }
-
-        getLoaderManager().initLoader(0, null, this);
-    }
-
-    private boolean mayRequestContacts() {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
-            return true;
-        }
-        if (checkSelfPermission(READ_CONTACTS) == PackageManager.PERMISSION_GRANTED) {
-            return true;
-        }
-        if (shouldShowRequestPermissionRationale(READ_CONTACTS)) {
-            Snackbar.make(mEmailView, R.string.permission_rationale, Snackbar.LENGTH_INDEFINITE)
-                    .setAction(android.R.string.ok, new View.OnClickListener() {
-                        @Override
-                        @TargetApi(Build.VERSION_CODES.M)
-                        public void onClick(View v) {
-                            requestPermissions(new String[]{READ_CONTACTS}, REQUEST_READ_CONTACTS);
-                        }
-                    });
-        } else {
-            requestPermissions(new String[]{READ_CONTACTS}, REQUEST_READ_CONTACTS);
-        }
-        return false;
-    }
-
-    /**
-     * Callback received when a permissions request has been completed.
-     */
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
-                                           @NonNull int[] grantResults) {
-        if (requestCode == REQUEST_READ_CONTACTS) {
-            if (grantResults.length == 1 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                populateAutoComplete();
-            }
-        }
-    }
-
-    /**
-     * Attempts to sign in or register the account specified by the login form.
-     * If there are form errors (invalid email, missing fields, etc.), the
-     * errors are presented and no actual login attempt is made.
-     */
-    private void attemptLogin() {
-        if (mAuthTask != null) {
-            return;
-        }
-
-        // Reset errors.
-        mEmailView.setError(null);
-        mPasswordView.setError(null);
-
-        // Store values at the time of the login attempt.
-        String email = mEmailView.getText().toString();
-        String password = mPasswordView.getText().toString();
-        String credit = mCreditView.getText().toString();
-
-        boolean cancel = false;
-        View focusView = null;
-
-        // Check for a valid password, if the user entered one.
-        if (!TextUtils.isEmpty(password) && !isPasswordValid(password)) {
-            mPasswordView.setError(getString(R.string.error_invalid_password));
-            focusView = mPasswordView;
-            cancel = true;
-        }
-
-        // Check for a valid email address.
-        if (TextUtils.isEmpty(email)) {
-            mEmailView.setError(getString(R.string.error_field_required));
-            focusView = mEmailView;
-            cancel = true;
-        } else if (!isEmailValid(email)) {
-            mEmailView.setError(getString(R.string.error_invalid_email));
-            focusView = mEmailView;
-            cancel = true;
-        }
-
-
-        if (cancel) {
-            // There was an error; don't attempt login and focus the first
-            // form field with an error.
-            focusView.requestFocus();
-        } else {
-            // Show a progress spinner, and kick off a background task to
-            // perform the user login attempt.
-            mAuthTask = new UserLoginTask(email, password);
-            // mAuthTask.execute((Void) null);
-        }
-
-        mEmail = email;
-        mPassword = password;
-        mCredit = credit;
-
-//        SharedPreferences.Editor edit = sharedPreferences.edit();
-//        edit.putString("email", email);
-//        edit.putString("password", password);
-    }
-
-    private boolean isEmailValid(String email) {
-        //TODO: Replace this with your own logic
-        //return email.contains("@gmail.com");
-        return true;
-    }
-
-    private boolean isPasswordValid(String password) {
-        //TODO: Replace this with your own logic
-        //return password.length() > 4;
-        return true;
-    }
-
-    @Override
-    public Loader<Cursor> onCreateLoader(int i, Bundle bundle) {
-        return new CursorLoader(this,
-            // Retrieve data rows for the device user's 'profile' contact.
-            Uri.withAppendedPath(ContactsContract.Profile.CONTENT_URI,
-                    ContactsContract.Contacts.Data.CONTENT_DIRECTORY), ProfileQuery.PROJECTION,
-
-            // Select only email addresses.
-            ContactsContract.Contacts.Data.MIMETYPE +
-                    " = ?", new String[]{ContactsContract.CommonDataKinds.Email
-            .CONTENT_ITEM_TYPE},
-
-            // Show primary email addresses first. Note that there won't be
-            // a primary email address if the user hasn't specified one.
-            ContactsContract.Contacts.Data.IS_PRIMARY + " DESC");
-    }
-
-    @Override
-    public void onLoadFinished(Loader<Cursor> cursorLoader, Cursor cursor) {
-        List<String> emails = new ArrayList<>();
-        cursor.moveToFirst();
-        while (!cursor.isAfterLast()) {
-            emails.add(cursor.getString(ProfileQuery.ADDRESS));
-            cursor.moveToNext();
-        }
-
-        addEmailsToAutoComplete(emails);
-    }
-
-    @Override
-    public void onLoaderReset(Loader<Cursor> cursorLoader) {
-
-    }
-
-    private void addEmailsToAutoComplete(List<String> emailAddressCollection) {
-        //Create adapter to tell the AutoCompleteTextView what to show in its dropdown list.
-        ArrayAdapter<String> adapter =
-                new ArrayAdapter<>(MainActivity.this,
-                        android.R.layout.simple_dropdown_item_1line, emailAddressCollection);
-
-        mEmailView.setAdapter(adapter);
-    }
-
-
-    private interface ProfileQuery {
-        String[] PROJECTION = {
-                ContactsContract.CommonDataKinds.Email.ADDRESS,
-                ContactsContract.CommonDataKinds.Email.IS_PRIMARY,
-        };
-
-        int ADDRESS = 0;
-        int IS_PRIMARY = 1;
-    }
-
-    /**
-     * Represents an asynchronous login/registration task used to authenticate
-     * the user.
-     */
-    public class UserLoginTask extends AsyncTask<Void, Void, Boolean> {
-
-        private final String mEmail;
-        private final String mPassword;
-
-        UserLoginTask(String email, String password) {
-            mEmail = email;
-            mPassword = password;
-        }
-
-        @Override
-        protected Boolean doInBackground(Void... params) {
-            // TODO: attempt authentication against a network service.
-
-            try {
-                // Simulate network access.
-                Thread.sleep(2000);
-            } catch (InterruptedException e) {
-                return false;
-            }
-
-            for (String credential : DUMMY_CREDENTIALS) {
-                String[] pieces = credential.split(":");
-                if (pieces[0].equals(mEmail)) {
-                    // Account exists, return true if the password matches.
-                    return pieces[1].equals(mPassword);
-                }
-            }
-
-            // TODO: register the new account here.
-            return true;
-        }
-
-        @Override
-        protected void onPostExecute(final Boolean success) {
-            mAuthTask = null;
-
-            if (success) {
-                finish();
-            } else {
-                mPasswordView.setError(getString(R.string.error_incorrect_password));
-                mPasswordView.requestFocus();
-            }
-        }
-
-        @Override
-        protected void onCancelled() {
-            mAuthTask = null;
-        }
     }
 }
-
