+++ TransnetAndroid/app/src/main/res/layout/offer_ride.xml
@@ -8,7 +8,7 @@
         android:layout_height="wrap_content"
         android:text="Where are you leaving from?"
         android:id="@+id/leavingLabel"
-        android:textSize="25dp"
+        android:textSize="20dp"
         android:textStyle="bold"
         android:layout_centerHorizontal="true"
         android:layout_margin="20dp"/>
@@ -28,7 +28,7 @@
         android:text="Where are you arriving to?"
         android:layout_below="@+id/place_autocomplete_fragment1"
         android:id="@+id/arrivingLabel"
-        android:textSize="25dp"
+        android:textSize="20dp"
         android:textStyle="bold"
         android:layout_centerHorizontal="true"
         android:layout_margin="20dp"/>
@@ -48,10 +48,10 @@
         android:text="When are you leaving?"
         android:layout_below="@+id/place_autocomplete_fragment2"
         android:id="@+id/timeLeavingLabel"
-        android:textSize="25dp"
+        android:textSize="20dp"
         android:textStyle="bold"
         android:layout_centerHorizontal="true"
-        android:layout_margin="20dp"/>
+        android:layout_margin="10dp"/>
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -115,10 +115,10 @@
         android:text="When are you arriving?"
         android:layout_below="@+id/timeLeavingLayout"
         android:id="@+id/timeArrivigLabelLabel"
-        android:textSize="25dp"
+        android:textSize="20dp"
         android:textStyle="bold"
         android:layout_centerHorizontal="true"
-        android:layout_margin="20dp"/>
+        android:layout_margin="10dp"/>
 
     <LinearLayout
         android:layout_width="match_parent"
